Remove long from type

+++ metnet/layers/ConditionTime.py
@@ -6,7 +6,7 @@ def condition_time(x, i=0, size=(12, 16), seq_len=15):
     "create one hot encoded time image-layers, i in [1, seq_len]"
     assert i < seq_len
     times = (
-        (torch.eye(seq_len, dtype=torch.long, device=x.device)[i.long()])
+        (torch.eye(seq_len, dtype=torch.long, device=x.device)[i])
         .unsqueeze(-1)
         .unsqueeze(-1)
     )
